feat(nix): init package
if you want to start this program in production environment,
first set the`SECRET_KEY_BASE` environment variable, then
use `nix run .# start --impure` to run it locally.

to start the program in dev environment, use `nix run .#dev start`

the only shortcoming is that the dependencies hash needs to be updated
manually.
It is possible to setup the CI to create a pull request with the
updated hash when `nix flake check` fails.

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>