feat(docker): init
Add Dockerfile for the Edgehog Device Forwarder.

The style is similar to Edgehog's Dockerfile,
but using alpine for reduced storage consumption.

By using alpine, running the elixir-alpine image
is required for the runner image too,
as using the standard alpine:latest image
the program won't run due to `erts` looking for
`glibc`.

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>